requirements.txt
onnxruntime
gdown
yacs
webdataset
scikit-image
torchgeometry>=0.1.2
PyYAML
configargparse
pillow
opencv-python
tqdm
smplx
pandas
xtcocotools
einops
mmcv==1.3.9
chumpy
mmhuman3d==0.11.0
tensorboard==2.14.0
tqdm==4.66.1
portalocker
loguru
configer
transforms3d
moviepy
numpy==1.23.5